fix(migrations)!: fix data type in DDL (replaced integer with bigint)

+++ src/main/resources/db/changelog/migrations/create-product-schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE oms_file_sources
 (
-    id               INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id               BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     file_name        VARCHAR NOT NULL,
     file_description VARCHAR,
     file_url         VARCHAR UNIQUE NOT NULL
@@ -8,22 +8,22 @@ CREATE TABLE oms_file_sources
 
 CREATE TABLE oms_types
 (
-    id               INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id               BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     type_name        VARCHAR UNIQUE NOT NULL,
     type_description VARCHAR UNIQUE NOT NULL
 );
 
 CREATE TABLE oms_products
 (
-    id          INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id          BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name        VARCHAR UNIQUE NOT NULL,
     description VARCHAR        NOT NULL
 );
 
 CREATE TABLE oms_product_type
 (
-    product_id INTEGER,
-    type_id    INTEGER,
+    product_id BIGINT,
+    type_id    BIGINT,
     CONSTRAINT fk_product_id
         FOREIGN KEY (product_id)
             REFERENCES oms_products (id),
@@ -34,8 +34,8 @@ CREATE TABLE oms_product_type
 
 CREATE TABLE oms_product_file_sources
 (
-    product_id INTEGER,
-    file_id INTEGER,
+    product_id BIGINT,
+    file_id BIGINT,
     CONSTRAINT fk_product_id
         FOREIGN KEY (product_id)
             REFERENCES oms_products (id),
